cat_dog_classifier/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow
streamlit
